Fix `implicit-function-declaration` warnings, pass 6

+++ include/sysdolphin/baselib/objalloc.h
@@ -63,6 +63,6 @@ s32 HSD_ObjAllocAddFree(HSD_ObjAllocData* data, u32 num);
 void* HSD_ObjAlloc(HSD_ObjAllocData* data);
 void HSD_ObjFree(HSD_ObjAllocData* data, void* obj);
 void _HSD_ObjAllocForgetMemory(void);
-void HSD_ObjAllocInit(HSD_ObjAllocData* data, u32 size, u32 align);
+void HSD_ObjAllocInit(HSD_ObjAllocData* data, size_t size, u32 align);
 
 #endif
